# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Blaaize Wallet API'
  description: 'This is the API documentation for Blaaize Wallet.'
  version: 1.0.0
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: 'Register a new user'
      operationId: d764dd091cc4494ae0baf360b03319f3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - firstname
                - lastname
                - email
                - password
              properties:
                firstname:
                  type: string
                  example: John
                lastname:
                  type: string
                  example: Doe
                email:
                  type: string
                  format: email
                  example: johndoe@example.com
                password:
                  type: string
                  format: password
                  example: StrongPassword123
              type: object
      responses:
        201:
          description: 'User registered successfully'
        422:
          description: 'Validation error'
        500:
          description: 'Server error'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: 'User login'
      operationId: 8dcb70df1020986038d098cc08d05dae
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: johndoe@example.com
                password:
                  type: string
                  format: password
                  example: StrongPassword123
              type: object
      responses:
        200:
          description: 'Login successful'
        401:
          description: 'Invalid credentials'
        422:
          description: 'Validation error'
  /api/wallet/transfer:
    post:
      tags:
        - Transaction
      summary: 'Transfer funds between wallets'
      description: "Transfers funds from one user's wallet to another, converting currency if needed."
      operationId: transferFunds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - sender_wallet_id
                - receiver_wallet_id
                - amount
                - from_currency
                - to_currency
              properties:
                sender_wallet_id:
                  description: "ID of the sender's wallet"
                  type: integer
                  example: 1
                receiver_wallet_id:
                  description: "ID of the receiver's wallet"
                  type: integer
                  example: 2
                amount:
                  description: 'Amount to transfer'
                  type: number
                  format: float
                  example: 100.5
                from_currency:
                  description: "Currency of the sender's wallet"
                  type: string
                  example: USD
                to_currency:
                  description: "Currency of the receiver's wallet"
                  type: string
                  example: EUR
                description:
                  description: 'Optional transaction description'
                  type: string
                  example: 'Payment for services'
              type: object
      responses:
        200:
          description: 'Transfer successful'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Transfer successful'
                  data:
                    properties:
                      transaction_reference:
                        type: string
                        example: 550e8400-e29b-41d4-a716-446655440000
                      sender_wallet_id:
                        type: integer
                        example: 1
                      receiver_wallet_id:
                        type: integer
                        example: 2
                      amount:
                        type: number
                        format: float
                        example: 100.5
                      converted_amount:
                        type: number
                        format: float
                        example: 95.75
                      currency:
                        type: string
                        example: USD
                      to_currency:
                        type: string
                        example: EUR
                      exchange_rate:
                        type: number
                        format: float
                        example: 0.9575
                      transaction_fee:
                        type: number
                        format: float
                        example: 0
                      status:
                        type: string
                        example: completed
                      description:
                        type: string
                        example: 'Payment for services'
                    type: object
                type: object
        400:
          description: 'Invalid request or insufficient balance'
        401:
          description: Unauthorized
        422:
          description: 'Validation failed'
        500:
          description: 'Internal Server Error'
      security:
        -
          sanctum: []
  /api/transfer-request/request:
    post:
      tags:
        - 'Transfer Requests'
      summary: 'Request money from another user'
      operationId: 070d81fa8c1bae615a5710ca60750523
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - recipient_id
                - request_wallet_id
                - amount
                - currency
              properties:
                recipient_id:
                  type: integer
                  example: 5
                request_wallet_id:
                  type: integer
                  example: 1
                amount:
                  type: number
                  format: float
                  example: 50
                currency:
                  type: string
                  example: USD
              type: object
      responses:
        200:
          description: 'Request placed successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Transfer request placed successfully'
                type: object
        422:
          description: 'Validation failed'
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
  '/api/transfer-request/accept-request/{id}':
    post:
      tags:
        - 'Transfer Requests'
      summary: 'Accept a money request'
      operationId: 06d625beb25653fd032edaafe345a886
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Request accepted and money transferred successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Money request accepted successfully. Funds have been transferred.'
                type: object
        404:
          description: 'Request not found'
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
  '/api/transfer-request/reject-request/{id}':
    post:
      tags:
        - 'Transfer Requests'
      summary: 'Reject a money request'
      operationId: aea979aa76cd440dc8a8ed654a67c5d9
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Request rejected successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Money request rejected successfully.'
                type: object
        404:
          description: 'Request not found'
        500:
          description: 'Server error'
      security:
        -
          sanctum: []
...
components:
  securitySchemes:
    sanctum:
      type: http
      bearerFormat: JWT
      scheme: bearer
...
security:
  -
    sanctum:
      type: apiKey
      description: 'Enter token in format (<token>)'
      name: Authorization
      in: header